+++ qtypes/__init__.py
@@ -2,5 +2,6 @@ from . import widgets
 from .__version__ import *
 from ._bool import *
 from ._enum import *
+from ._filepath import *
 from ._number import *
 from ._string import *