+++ qtypes/_string.py
@@ -14,9 +14,9 @@ class String(Base):
     def give_control(self, control_widget):
         self.widget = control_widget
         # fill out items
-        self.widget.setText(str(self.value.read()))
+        self.widget.setText(str(self.value.get()))
         # connect signals and slots
-        self.updated.connect(lambda: self.widget.setText(self.value.read()))
-        self.widget.editingFinished.connect(lambda: self.write(str(self.widget.text())))
+        self.updated.connect(lambda: self.widget.setText(self.value.get()))
+        self.widget.editingFinished.connect(lambda: self.set(str(self.widget.text())))
         self.widget.setToolTip(self.tool_tip)
         self.has_widget = True